Multitude of Python 2 compatiblity fixes.
I think targeting Python 2.7 and then working to get it running on
Python 3 is silly. I prefer targeting Python 3 first, then seeing if
it'll work in Python 2 without major changes. Python 3 is where new work
is happening, and I want to support the language by helping to move it
forward.

The biggest change is that I'm explicitly closing a bunch of url
requests now.